+++ GANDLF/data/augmentation/__init__.py
@@ -16,7 +16,8 @@ from .rotations import (
     rotate_90,
     rotate_180,
 )
-from .rgb_augs import colorjitter_transform, hed_transform
+from .rgb_augs import colorjitter_transform
+from .hed_augs import hed_transform
 
 # Defining a dictionary for augmentations - key is the string and the value is the augmentation object
 global_augs_dict = {
